Fixing access of internal resources bundle
