+++ Adyen/Utilities/Localization.swift
@@ -93,21 +93,3 @@ public func ADYLocalizedSubmitButtonTitle(with amount: Payment.Amount?, _ parame
     
     return ADYLocalizedString("adyen.submitButton.formatted", parameters, formattedAmount)
 }
-
-internal extension Bundle {
-    // swiftlint:disable explicit_acl
-    
-    /// The main bundle of the framework.
-    static let core: Bundle = {
-        Bundle(for: Coder.self)
-    }()
-    
-    /// The bundle in which the framework's resources are located.
-    static let internalResources: Bundle = {
-        let url = core.url(forResource: "Adyen", withExtension: "bundle")
-        let bundle = url.flatMap { Bundle(url: $0) }
-        return bundle ?? core
-    }()
-    
-    // swiftlint:enable explicit_acl
-}
